Add ARG for CLI_URL
Pulling the URL out into an ARG makes it clearer to bump the CLI versions. You could just have the Version number in the ARG, but I find it easier to take the whole URL from GitHub and is less prone to mistakes if changes are made to the Balena repo.